feat(migration): add legacy user migration and run on launch
There is a chance that the save extension will be launched before the main app, so we have to attempt to run the migration at both the entry point to the main app, as well as the save extension. Since the save extension doesn't have an equivalent of an app delegate, we use the view controller used on open to perform the migration.